refactor: colocate the folder dashboard page under src/page/FolderPage
- Move the page and components from src/scenes/Folder, matching the page-colocation convention
- The page is plain React rather than scenes-based, so it no longer belongs under src/scenes
- Export FolderPage through an index for the router import
