+++ src/page/FolderPage/index.ts
@@ -1,0 +1,1 @@
+export { FolderPage } from './FolderPage';
